Splitting ConfigurableAreas and TableAttributes into Traits
