+++ src/Traits/ComponentUtilities.php
@@ -32,47 +32,11 @@ trait ComponentUtilities
 
     protected bool $eagerLoadAllRelationsStatus = false;
 
-    protected array $componentWrapperAttributes = [];
-
-    protected array $tableWrapperAttributes = [];
-
-    protected array $tableAttributes = [];
-
-    protected array $theadAttributes = [];
-
-    protected array $tbodyAttributes = [];
-
-    protected $thAttributesCallback;
-
-    protected $thSortButtonAttributesCallback;
-
-    protected $trAttributesCallback;
-
-    protected $trUrlCallback;
-
-    protected $trUrlTargetCallback;
-
-    protected $tdAttributesCallback;
-
     protected bool $collapsingColumnsStatus = true;
 
     protected string $emptyMessage = 'No items found. Try to broaden your search.';
 
     protected array $additionalSelects = [];
-
-    protected bool $hideConfigurableAreasWhenReorderingStatus = true;
-
-    protected array $configurableAreas = [
-        'before-tools' => null,
-        'toolbar-left-start' => null,
-        'toolbar-left-end' => null,
-        'toolbar-right-start' => null,
-        'toolbar-right-end' => null,
-        'before-toolbar' => null,
-        'after-toolbar' => null,
-        'before-pagination' => null,
-        'after-pagination' => null,
-    ];
 
     // Sets the Theme If Not Already Set
     public function mountComponentUtilities(): void
