+++ src/Traits/Helpers/ComponentHelpers.php
@@ -92,78 +92,6 @@ trait ComponentHelpers
     }
 
     /**
-     * @return array<mixed>
-     */
-    public function getComponentWrapperAttributes(): array
-    {
-        return count($this->componentWrapperAttributes) ? $this->componentWrapperAttributes : ['id' => 'datatable-'.$this->getId()];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getTableWrapperAttributes(): array
-    {
-        return count($this->tableWrapperAttributes) ? $this->tableWrapperAttributes : ['default' => true];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getTableAttributes(): array
-    {
-        return count($this->tableAttributes) ? $this->tableAttributes : ['id' => 'table-'.$this->getTableName(), 'default' => true];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getTheadAttributes(): array
-    {
-        return count($this->theadAttributes) ? $this->theadAttributes : ['default' => true];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getTbodyAttributes(): array
-    {
-        return count($this->tbodyAttributes) ? $this->tbodyAttributes : ['default' => true];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getThAttributes(Column $column): array
-    {
-        return $this->thAttributesCallback ? call_user_func($this->thAttributesCallback, $column) : ['default' => true];
-    }
-
-    /**
-     * @return  array<mixed>
-     */
-    public function getThSortButtonAttributes(Column $column): array
-    {
-        return $this->thSortButtonAttributesCallback ? call_user_func($this->thSortButtonAttributesCallback, $column) : ['default' => true];
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function getTrAttributes(Model $row, int $index): array
-    {
-        return $this->trAttributesCallback ? call_user_func($this->trAttributesCallback, $row, $index) : ['default' => true];
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function getTdAttributes(Column $column, Model $row, int $colIndex, int $rowIndex): array
-    {
-        return $this->tdAttributesCallback ? call_user_func($this->tdAttributesCallback, $column, $row, $colIndex, $rowIndex) : ['default' => true];
-    }
-
-    /**
      * Get the translated empty message of the table
      */
     public function getEmptyMessage(): string
@@ -236,21 +164,6 @@ trait ComponentHelpers
         return $this->getCollapsingColumnsStatus() === false;
     }
 
-    public function hasTableRowUrl(): bool
-    {
-        return $this->trUrlCallback !== null;
-    }
-
-    public function getTableRowUrl($row): ?string
-    {
-        return $this->trUrlCallback ? call_user_func($this->trUrlCallback, $row) : null;
-    }
-
-    public function getTableRowUrlTarget($row): ?string
-    {
-        return $this->trUrlTargetCallback ? call_user_func($this->trUrlTargetCallback, $row) : null;
-    }
-
     /**
      * @return array<mixed>
      */
@@ -259,64 +172,4 @@ trait ComponentHelpers
         return $this->additionalSelects;
     }
 
-    /**
-     * @return array<mixed>
-     */
-    public function getConfigurableAreas(): array
-    {
-        return $this->configurableAreas;
-    }
 
-    public function hasConfigurableAreaFor(string $area): bool
-    {
-        if ($this->hideConfigurableAreasWhenReorderingIsEnabled() && $this->reorderIsEnabled() && $this->currentlyReorderingIsEnabled()) {
-            return false;
-        }
-
-        return isset($this->configurableAreas[$area]) && $this->getConfigurableAreaFor($area) !== null;
-    }
-
-    /**
-     * @param  string|array<mixed>  $area
-     */
-    public function getConfigurableAreaFor($area): ?string
-    {
-        $area = $this->configurableAreas[$area] ?? null;
-
-        if (is_array($area)) {
-            return $area[0];
-        }
-
-        return $area;
-    }
-
-    /**
-     * @param  string|array<mixed>  $area
-     * @return array<mixed>
-     */
-    public function getParametersForConfigurableArea($area): array
-    {
-        $area = $this->configurableAreas[$area] ?? null;
-
-        if (is_array($area) && isset($area[1]) && is_array($area[1])) {
-            return $area[1];
-        }
-
-        return [];
-    }
-
-    public function getHideConfigurableAreasWhenReorderingStatus(): bool
-    {
-        return $this->hideConfigurableAreasWhenReorderingStatus;
-    }
-
-    public function hideConfigurableAreasWhenReorderingIsEnabled(): bool
-    {
-        return $this->getHideConfigurableAreasWhenReorderingStatus() === true;
-    }
-
-    public function hideConfigurableAreasWhenReorderingIsDisabled(): bool
-    {
-        return $this->getHideConfigurableAreasWhenReorderingStatus() === false;
-    }
-}
